Import of CREAMYG/KinoSearch-0.30_04 from CPAN.
gitpan-cpan-distribution: KinoSearch
gitpan-cpan-version:      0.30_04
gitpan-cpan-path:         CREAMYG/KinoSearch-0.30_04.tar.gz
gitpan-cpan-author:       CREAMYG
gitpan-cpan-maturity:     developer

+++ core/KinoSearch/Index/HighlightReader.c
@@ -210,7 +210,7 @@ DefHLReader_read_record(DefaultHighlightReader *self, i32_t doc_id,
         /* Copy the whole record. */
         u64_t  filepos = InStream_Read_U64(ix_in);
         u64_t  end     = InStream_Read_U64(ix_in);
-        size_t size    = end - filepos;
+        size_t size    = (size_t)(end - filepos);
         BB_Grow(target, size);
         InStream_Seek(dat_in, filepos);
         InStream_Read_Bytes(dat_in, target->ptr, size);
